# openapi.json
{
    "openapi": "3.0.2",
    "info": {
    "title": "FastAPI",
    "version": "0.1.0"
    },
    "paths": {
    "/blog": {
    "get": {
    "summary": "All",
    "operationId": "all_blog_get",
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "title": "Response All Blog Get",
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/ShowBlog"
    }
    }
    }
    }
    }
    }
    },
    "post": {
    "summary": "Create",
    "operationId": "create_blog_post",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Blog"
    }
    }
    },
    "required": true
    },
    "responses": {
    "201": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/blog/{id}": {
    "get": {
    "summary": "Show",
    "operationId": "show_blog__id__get",
    "parameters": [
    {
    "required": true,
    "schema": {
    "title": "Id"
    },
    "name": "id",
    "in": "path"
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ShowBlog"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "put": {
    "summary": "Update",
    "operationId": "update_blog__id__put",
    "parameters": [
    {
    "required": true,
    "schema": {
    "title": "Id"
    },
    "name": "id",
    "in": "path"
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Blog"
    }
    }
    },
    "required": true
    },
    "responses": {
    "202": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "delete": {
    "summary": "Destroy",
    "operationId": "destroy_blog__id__delete",
    "parameters": [
    {
    "required": true,
    "schema": {
    "title": "Id"
    },
    "name": "id",
    "in": "path"
    }
    ],
    "responses": {
    "204": {
    "description": "Successful Response"
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/user": {
    "post": {
    "summary": "Create User",
    "operationId": "create_user_user_post",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/User"
    }
    }
    },
    "required": true
    },
    "responses": {
    "201": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ShowUser"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/use/{id}": {
    "get": {
    "summary": "Get User",
    "operationId": "get_user_use__id__get",
    "parameters": [
    {
    "required": true,
    "schema": {
    "title": "Id",
    "type": "integer"
    },
    "name": "id",
    "in": "path"
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ShowUser"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    }
    },
    "components": {
    "schemas": {
    "Blog": {
    "title": "Blog",
    "required": [
    "title",
    "body"
    ],
    "type": "object",
    "properties": {
    "title": {
    "title": "Title",
    "type": "string"
    },
    "body": {
    "title": "Body",
    "type": "string"
    }
    }
    },
    "HTTPValidationError": {
    "title": "HTTPValidationError",
    "type": "object",
    "properties": {
    "detail": {
    "title": "Detail",
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/ValidationError"
    }
    }
    }
    },
    "ShowBlog": {
    "title": "ShowBlog",
    "required": [
    "title",
    "body",
    "id"
    ],
    "type": "object",
    "properties": {
    "title": {
    "title": "Title",
    "type": "string"
    },
    "body": {
    "title": "Body",
    "type": "string"
    },
    "id": {
    "title": "Id",
    "type": "integer"
    }
    }
    },
    "ShowUser": {
    "title": "ShowUser",
    "required": [
    "name",
    "email",
    "id"
    ],
    "type": "object",
    "properties": {
    "name": {
    "title": "Name",
    "type": "string"
    },
    "email": {
    "title": "Email",
    "type": "string"
    },
    "id": {
    "title": "Id",
    "type": "integer"
    }
    }
    },
    "User": {
    "title": "User",
    "required": [
    "name",
    "email",
    "password"
    ],
    "type": "object",
    "properties": {
    "name": {
    "title": "Name",
    "type": "string"
    },
    "email": {
    "title": "Email",
    "type": "string"
    },
    "password": {
    "title": "Password",
    "type": "string"
    }
    }
    },
    "ValidationError": {
    "title": "ValidationError",
    "required": [
    "loc",
    "msg",
    "type"
    ],
    "type": "object",
    "properties": {
    "loc": {
    "title": "Location",
    "type": "array",
    "items": {
    "type": "string"
    }
    },
    "msg": {
    "title": "Message",
    "type": "string"
    },
    "type": {
    "title": "Error Type",
    "type": "string"
    }
    }
    }
    }
    }
    }